:recycle: extract class from strategy
